step9/default: 리팩토링
